Add a feature to enable/disable audio as needed
To enable the audio thread set enable_audio to 1 under general
settings in the JSON settings file.

The audio thread is responsible for capturing audio frames from
the microphone and feeding it to the V4L2 device. However due to
some race conditions sometimes this doesn't work and the video
doesn't come up on a reboot. Adding this option so at least
it can disabled until a proper fix is found.